requirements.txt
Django==4.0.4
python-decouple==3.6
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6